django_prj/requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-extensions==3.1.5
django-rest-framework==0.1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
flower==1.0.0
humanize==4.1.0
kombu==5.2.4
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
